# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register new account request (pending admin approval)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login with username/password and set auth cookies",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Logout current session and clear auth cookies",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Logout all sessions for current user and clear auth cookies",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/users": {
      "get": {
        "operationId": "AuthController_users",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List all users for admin management",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/approve": {
      "patch": {
        "operationId": "AuthController_approveUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Approve user registration request (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/activate": {
      "patch": {
        "operationId": "AuthController_activateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Activate user account (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/deactivate": {
      "patch": {
        "operationId": "AuthController_deactivateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Deactivate user account (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/reset-password": {
      "patch": {
        "operationId": "AuthController_resetUserPassword",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reset user password and revoke active sessions (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/sessions": {
      "get": {
        "operationId": "AuthController_userSessions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List active sessions for selected user (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/users/{userId}/sessions/{sessionId}": {
      "delete": {
        "operationId": "AuthController_revokeUserSession",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Revoke selected user session (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sessions": {
      "get": {
        "operationId": "AuthController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List active sessions for current admin account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sessions/{sessionId}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Revoke a specific active session for current admin account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me/sessions": {
      "get": {
        "operationId": "AuthController_mySessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List active sessions for current authenticated account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me/sessions/{sessionId}": {
      "delete": {
        "operationId": "AuthController_revokeMySession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Revoke selected active session for current authenticated account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Change password for current authenticated account and revoke other sessions",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SwimSync API",
    "description": "REST API for SwimSync services",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      }
    },
    "schemas": {
...
      "RegisterUserDto": {
        "type": "object",
        "properties": {}
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {}
      },
      "ResetUserPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}